Added features for Lend Transaction

+++ src/main/java/com/educare/unitylend/model/Wallet.java
@@ -12,5 +12,5 @@ import java.util.UUID;
 public class Wallet {
     private String walletId;
     private User user;
-    private Float balance;
+    private BigDecimal balance;
 }
